refactor(bridge): delete orphaned gascity cluster, keep live semver/codex helpers (ag-hfc #gascity-orphan-cluster)
3.1 teardown S2. The gascity compat was orphaned: agentworker's GasCityClientAdapter/NewGasCityWorker had ZERO callers repo-wide, internal/gascity was imported only by that dead adapter, and bridge/gc.go's GC* symbols were used only by gascity/fallback.go.

CORRECTS the stale 'pure rm / doctor-hint importers' premise (closed ag-2oi): bridge/gc.go was MIXED — it also held CompareSemver/ParseSemverParts, used LIVE by cmd/ao/codex_runtime.go. Extracted those to bridge/semver.go before deleting gc.go. The doctor gc-detection (fix_bridges.go) is independent (shells out to gc binary, imports neither pkg) — no rewire needed.

Deleted: internal/gascity (whole pkg), agentworker/gascity.go + gascity_test.go, bridge/gc.go + its 12 gascity test fns. Preserved: bridge codex helpers, agentworker.QuarantineWriter, and agentworker firstNonEmpty helper (gascity.go had defined it; relocated to agentworker/strings.go — build caught the cross-file coupling).

build+vet green, 76 bridge+agentworker tests pass, full go test ./... green. Updated stale CLAUDE.md gascity note.
