--- cli/internal/gascity/testdata/city-response.json
@@ -1,5 +1,0 @@
-{
-  "ok": true,
-  "name": "agentops",
-  "path": "/tmp/agentops-city"
-}
