--- cli/internal/gascity/testdata/health.json
@@ -1,4 +1,0 @@
-{
-  "ok": true,
-  "status": "ok"
-}
