--- cli/internal/gascity/testdata/heartbeat.json
@@ -1,3 +1,0 @@
-{
-  "timestamp": "2026-04-28T20:45:10Z"
-}
