--- cli/internal/gascity/testdata/problem-details.json
@@ -1,7 +1,0 @@
-{
-  "type": "about:blank",
-  "title": "Forbidden",
-  "status": 403,
-  "detail": "csrf: X-GC-Request header required on mutation endpoints",
-  "instance": "/v0/city/agentops/sessions"
-}
