--- cli/internal/gascity/testdata/session-submit-request.json
@@ -1,4 +1,0 @@
-{
-  "message": "continue the current RPI phase",
-  "intent": "follow_up"
-}
