--- cli/internal/gascity/testdata/transcript.json
@@ -1,16 +1,0 @@
-{
-  "id": "sess_123",
-  "template": "agentops/codex",
-  "provider": "codex",
-  "format": "conversation",
-  "turns": [
-    {
-      "role": "user",
-      "text": "continue"
-    },
-    {
-      "role": "assistant",
-      "text": "done"
-    }
-  ]
-}
